+++ include/str.h
@@ -8,7 +8,7 @@
 
 #define Str_acatv(s, ...) Str_acatvx(s, __VA_ARGS__, NULL)
 
-extern char*    Str_asub    (const char *s, int i, int j);
+extern char*    Str_asub    (const char *s, size_t i, size_t j);
 extern char*    Str_adup    (const char *s);
 extern char*    Str_acat    (const char *s1, const char *s2);
 extern char*    Str_acatvx  (const char *s, ...);
@@ -17,7 +17,7 @@ extern char*    Str_amap    (const char *s, const char *from, const char *to);
 
 extern char*    Str_asprintf(const char *fmt, ...);
 
-extern char**   Str_split(char* s, const char* delimiters, int empties);
+extern char**   Str_split(char* s, const char* delimiters, unsigned empties);
 
 extern uint32_t* Str_atoucs   (const char *src);
 extern char*     Str_atoutf8   (const uint32_t *src);
